Bump file_picker_writable
Fixes directory resolution going stale; see
https://github.com/hpoul/file_picker_writable/commit/a845daf3b51da7d9251530597eea37621b7dc45c

The problem seems to be that access to a directory is "remembered" for a while,
but then will "expire". After expiration, when bookmarking a URL under that
directory, you need to call startAccessingSecurityScopedResource() on the
parent, or else access will be denied.

Previously things seemed to be working because during the initial period when
access is "remembered", you don't need to call
startAccessingSecurityScopedResource() and it would still work.